requirements.txt
streamlit
google-genai
Pillow
pandas
streamlit-google-sheets